+++ services/spotifyService.ts
@@ -4,14 +4,12 @@ import myFetch from './http';
 export const spotifyService = {
   checkTokenAvailable: () => myFetch.get('/token'),
 
-  postSearchArtist: (artist: string): Promise<Artist[]> => {
-    return myFetch.post('/artists', {
-      method: 'POST',
-      body: artist,
-    });
+  fetchSearchArtistResult: (artist: string): Promise<Artist[]> | [] => {
+    if (!artist) return [];
+    return myFetch.get(`/artists?name=${artist}`);
   },
 
   fetchRelatedArtistGraphData: (id: string): Promise<Network> => {
-    return myFetch.get(`/artists?id=${id}`);
+    return myFetch.get(`/artists/${id}`);
   },
 };
